+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/tags/CSMTextSettingsTag.java
@@ -127,7 +127,7 @@ public class CSMTextSettingsTag extends Tag implements CharacterIdTag {
     }
 
     @Override
-    public void getNeededCharacters(Set<Integer> needed) {
+    public void getNeededCharacters(Set<Integer> needed, SWF swf) {
         needed.add(textID);
     }   
 }