+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/tags/DefineButtonSoundTag.java
@@ -142,7 +142,7 @@ public class DefineButtonSoundTag extends Tag implements CharacterIdTag {
     }
 
     @Override
-    public void getNeededCharacters(Set<Integer> needed) {
+    public void getNeededCharacters(Set<Integer> needed, SWF swf) {
         if (buttonSoundChar0 != 0) {
             needed.add(buttonSoundChar0);
         }