+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/tags/DefineFontAlignZonesTag.java
@@ -119,7 +119,7 @@ public class DefineFontAlignZonesTag extends Tag implements CharacterIdTag {
     }
     
     @Override
-    public void getNeededCharacters(Set<Integer> needed) {
+    public void getNeededCharacters(Set<Integer> needed, SWF swf) {
         needed.add(fontID);
     }
 }