+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/tags/DefineScalingGridTag.java
@@ -217,7 +217,7 @@ public class DefineScalingGridTag extends Tag implements CharacterIdTag {
     }
     
     @Override
-    public void getNeededCharacters(Set<Integer> needed) {
+    public void getNeededCharacters(Set<Integer> needed, SWF swf) {
         needed.add(characterId);
     }