+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/tags/NameCharacterTag.java
@@ -94,7 +94,7 @@ public class NameCharacterTag extends Tag implements CharacterIdTag {
     }
     
     @Override
-    public void getNeededCharacters(Set<Integer> needed) {
+    public void getNeededCharacters(Set<Integer> needed, SWF swf) {
         needed.add(characterId);
     }
 }