+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/types/FILLSTYLE.java
@@ -16,8 +16,11 @@
  */
 package com.jpexs.decompiler.flash.types;
 
+import com.jpexs.decompiler.flash.SWF;
 import com.jpexs.decompiler.flash.tags.DefineShape3Tag;
 import com.jpexs.decompiler.flash.tags.DefineShape4Tag;
+import com.jpexs.decompiler.flash.tags.base.CharacterTag;
+import com.jpexs.decompiler.flash.tags.base.ImageTag;
 import com.jpexs.decompiler.flash.tags.base.NeedsCharacters;
 import com.jpexs.decompiler.flash.types.annotations.Conditional;
 import com.jpexs.decompiler.flash.types.annotations.ConditionalType;
@@ -86,7 +89,7 @@ public class FILLSTYLE implements NeedsCharacters, FieldChangeObserver, Serializ
     public MATRIX bitmapMatrix;
 
     @Override
-    public void getNeededCharacters(Set<Integer> needed) {
+    public void getNeededCharacters(Set<Integer> needed, SWF swf) {
         if ((fillStyleType == REPEATING_BITMAP)
                 || (fillStyleType == CLIPPED_BITMAP)
                 || (fillStyleType == NON_SMOOTHED_REPEATING_BITMAP)
@@ -94,7 +97,10 @@ public class FILLSTYLE implements NeedsCharacters, FieldChangeObserver, Serializ
             if (bitmapId == 65535) { //In some cases, this special value is used, but is not used. Ignore it. (#1851)
                 return;
             }
-            needed.add(bitmapId);
+            CharacterTag character = swf.getCharacter(bitmapId);
+            if (character instanceof ImageTag) {
+                needed.add(bitmapId);
+            }
         }
     }